Move DataflowTaskExplorer and associated classes to server core
* Remove aggregate-task module
* Rename CompositeTaskExplorer to DataflowTaskExplorer
