--- spring-cloud-dataflow-aggregate-task/README.adoc
@@ -1,10 +1,0 @@
-= Spring Cloud Dataflow Aggregate Task Module
-
-Spring Cloud Task and Spring Batch utilize a series of database tables to support storing data about Boot Application executions as well as Job executions.
-For each major release of these projects, their database schemas  adjust to meet the needs for the latest release.
-SCDF supports applications that may use the current release of these projects as well as a previous release.
-The `spring-cloud-dataflow-aggregate-task` module provides support for dataflow to query and mutate data in each of the schema versions.
-
-== Tests
-
-The tests for this module are located in the `spring-cloud-dataflow-server` module
