--- spring-cloud-dataflow-aggregate-task/src/test/resources/logback-test.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-	<include resource="org/springframework/boot/logging/logback/base.xml" />
-	<logger name="org.testcontainers" level="ERROR"/>
-	<logger name="com.github.dockerjava" level="ERROR"/>
-</configuration>
