+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/task/DataflowTaskExplorer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.cloud.dataflow.composite.task;
+package org.springframework.cloud.dataflow.server.task;
 
 import java.util.Collection;
 import java.util.Date;
@@ -29,7 +29,7 @@ import org.springframework.data.domain.Pageable;
  *
  * @author Corneil du Plessis
  */
-public interface CompositeTaskExplorer {
+public interface DataflowTaskExplorer {
 	/**
 	 * find a task execution given an execution id and schema target.
 	 *
