+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/task/TaskDefinitionReader.java
@@ -1,4 +1,4 @@
-package org.springframework.cloud.dataflow.composite.task;
+package org.springframework.cloud.dataflow.server.task;
 
 import org.springframework.cloud.dataflow.core.TaskDefinition;
